requirements.txt
snowflake-snowpark-python
snowflake
pyyaml
pandas >=1.0.0
jupyterlab
pyspark 
-e .
